修改lcd15相关名称
