--- tft/lcd145-st7789/tft
@@ -1,3 +1,0 @@
-TFT_DTS_NAME=tft145
-TFT_MODULE=( "wpi_fbtft"  "wpi_fb_st7789"  )
-USE_CONF_FILE=( "99-fbdev.conf" )
