+++ tft/lcd15-st7789/dtb/sun50i-h616-tft15.dts
@@ -11,7 +11,7 @@
    			#size-cells = <0>;
 			status = "okay";
 			lcd@0{
-				compatible = "walnutpi,lcd145_st7789";
+				compatible = "walnutpi,lcd15_st7789";
 				reg = <0>;
 				spi-max-frequency = <30000000>;
 				status = "okay";
@@ -20,7 +20,7 @@
 				reset-gpios = <&pio 2 15 1>; /* PC15  GPIO_ACTIVE_LOW */
 				width = <240>;
 				hight = <240>;
-				rotate = <180>;
+				rotate = <270>;
 				buswidth = <8>;
 				regwidth = <8>;
 				fps = <30>;
